refactor: separação da toolbar para uma classe separada
